Comment out button group in detail pages
After the wiki site was moved from reference.digilentinc.com to digilent.com/reference, the back button here broke, targeting digilent.com/doku.php/ instead of digilent.com/reference/doku.php, causing potential 404 errors. It seems that DOKU_REL is not working properly, as doku.php is being treated as relative to the base URL of the entire site, and is not using the dokuwiki configuration, which correctly specifies digilent.com/reference. This fix is a temporary workaround, and we should instead remove detail pages entirely, or fix the 404 error by modifying the link target.